Move _generate_sbom to vulnscan cli, avoid strong cycles
