+++ src/vulnxscan/utils.py
@@ -12,13 +12,10 @@ Utility functions
 
 
 import json
-import logging
-import pathlib
 import re
 import time
 import urllib.parse
 
-from tempfile import NamedTemporaryFile
 import pandas as pd
 
 from common.utils import (
@@ -34,7 +31,6 @@ from repology.exceptions import RepologyNoMatchingPackages
 from repology.repology_cli import Repology
 from repology.repology_cli import getargs as cli_getargs
 from repology.repology_cve import query_cve
-from sbomnix.sbomdb import SbomDb
 
 
 ################################################################################
@@ -209,22 +205,6 @@ def _github_query(query_str, delay=60):
     resp_json = json.loads(resp.text)
     LOG.log(LOG_SPAM, "total_count=%s", resp_json["total_count"])
     return resp_json
-
-
-def _generate_sbom(target_path, buildtime=False):
-    LOG.info("Generating SBOM for target '%s'", target_path)
-    sbomdb = SbomDb(target_path, buildtime, include_meta=False)
-    prefix = "vulnxscan_"
-    cdx_suffix = ".json"
-    csv_suffix = ".csv"
-    with NamedTemporaryFile(
-        delete=False, prefix=prefix, suffix=cdx_suffix
-    ) as fcdx, NamedTemporaryFile(
-        delete=False, prefix=prefix, suffix=csv_suffix
-    ) as fcsv:
-        sbomdb.to_cdx(fcdx.name, printinfo=False)
-        sbomdb.to_csv(fcsv.name, loglevel=logging.DEBUG)
-        return pathlib.Path(fcdx.name), pathlib.Path(fcsv.name)
 
 
 def _is_json(path):
